Move weapon-range-template to canvas module
