+++ src/lancer.ts
@@ -27,7 +27,7 @@ import { LancerPilotSheet } from "./module/actor/pilot-sheet";
 import { LancerFrameSheet } from "./module/item/frame-sheet";
 import { LancerItemSheet } from "./module/item/item-sheet";
 import { LancerLicenseSheet } from "./module/item/license-sheet";
-import { WeaponRangeTemplate } from "./module/pixi/weapon-range-template";
+import { WeaponRangeTemplate } from "./module/canvas/weapon-range-template";
 
 // Import helpers
 import { LCPManager, addLCPManagerButton } from "./module/apps/lcp-manager/lcp-manager";
