+++ src/module/canvas/weapon-range-template.ts
